backend/requirements.txt
aiosmtplib==3.0.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
astroid==3.3.9
autopep8==2.0.4
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docstring-to-markdown==0.15
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
fastapi-mail==1.4.2
flake8==7.1.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.7.0
openai==1.65.4
orjson==3.10.15
packaging==24.2
parso==0.8.4
passlib==1.7.4
platformdirs==4.3.6
pluggy==1.5.0
psycopg2==2.9.10
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
pymongo==4.11.2
python-dotenv==1.0.1
python-jose==3.4.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.2
python-multipart==0.0.20
pytoolconfig==1.3.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rope==1.13.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
sqladmin==0.20.1
SQLAlchemy==2.0.38
starlette==0.46.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
websockets==15.0
whatthepatch==1.0.7
WTForms==3.1.2
yapf==0.43.0
